Controller 구조 정리
