+++ holidaysomething/holiday-something-react/src/pages/Home.js
@@ -31,7 +31,7 @@ class Home extends Component {
 
   _callApi = () => {
     // 현재 로그인 되어 있는 user의 id를 가져와서 사용하는 것으로 변경해야 한다.
-    // return fetch('/api/userTmp?id=11')
+    // return fetch('/api/user/11')
     return fetch('/api/user/authenticated')  // 현재 로그인된 유저
     .then(response => response.json())
     .then(json => json)
