+++ holidaysomething/holiday-something-react/src/shared/App.js
@@ -23,7 +23,7 @@ class App extends Component {
 
   _callApi = () => {
     // 현재 로그인 되어 있는 user의 id를 가져와서 사용하는 것으로 변경해야 한다.
-    // return fetch('http://localhost:8080/api/userTmp?id=11')
+    // return fetch('http://localhost:8080/api/user/11')
     return fetch('/api/user/authenticated')  // 현재 로그인된 유저
     .then(response => response.json())
     .then(json => json)
